an interm update for finder.py
In this commit I put in the code to research what characters come after
the commands. I expect to delete this code in the next commit, as it
does not need to exist forever, but I dont want to have to rewrite it if
we ever need to look into this again.